+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## [Unreleased]
 ### Added
+- Add an inspection for duplicating the field value of the `Source` paragraph
+  in a `Package` paragraph of `debian/control`.
 
 ### Changed